refactoring(category) : [enrollment] CourseFinderAdapter 코드 수정
Category 코드 리팩토링에 의해 영향을 받는 [Enrollment] CourseFinderAdapter 클래스 일부 수정
