+++ src/main/java/com/lxp/aplus/category/application/internal/dto/CategoryInternalResult.java
