+++ src/main/java/com/lxp/aplus/category/application/internal/usecase/CategoryInternalUseCase.java
@@ -1,8 +1,7 @@
 package com.lxp.aplus.category.application.internal.usecase;
 
-import com.lxp.aplus.category.application.internal.dto.CategoryInternalDto;
+import com.lxp.aplus.category.application.internal.dto.CategoryInternalResult;
 import com.lxp.aplus.category.application.port.out.CategoryRepository;
-import com.lxp.aplus.category.domain.Category;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -13,8 +12,8 @@ import java.util.Optional;
 public class CategoryInternalUseCase {
     private final CategoryRepository categoryRepository;
 
-    public Optional<CategoryInternalDto> findByIdWithParent(Long id){
+    public Optional<CategoryInternalResult> findByIdWithParent(Long id){
         return categoryRepository.findByIdWithParent(id)
-                .map(CategoryInternalDto::from);
+                .map(CategoryInternalResult::from);
     }
 }
